run_make_support: move `impl_common_helpers` into own `macros` module
